:sparkles: feat(custom): 图片存入磁盘后写入数据库中,并通过接口获取数据库用户信息

+++ afterEnd/common/common-util/src/main/java/com/lxh/utils/token/UserToken.java
@@ -9,9 +9,9 @@ public class UserToken {
                 '}';
     }
 
-    private Long id;
+    private int id;
     private String username;
-    public UserToken setUserInfo(Long uid, String username) {
+    public UserToken setUserInfo(int uid, String username) {
         this.id = uid;
         this.username = username;
         return this;
@@ -19,7 +19,7 @@ public class UserToken {
     public String getUsername() {
         return username;
     }
-    public Long getId() {
+    public int getId() {
         return id;
     }
 }
